requirements.txt
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.1
selenium==4.43.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
